refactor: rename telemetry to usage report
