+++ src/schemas/appSettings.ts
@@ -65,7 +65,7 @@ const RecorderBrowserPathSchema = RecorderDetectBrowserPathSchema.extend({
   browserPath: z.string().optional(),
 })
 
-export const TelemetrySettingsSchema = z.object({
+export const UsageReportSettingsSchema = z.object({
   enabled: z.boolean(),
 })
 
@@ -98,5 +98,5 @@ export const AppSettingsSchema = z.object({
   proxy: ProxySettingsSchema,
   recorder: RecorderSettingsSchema,
   windowState: WindowStateSchema,
-  telemetry: TelemetrySettingsSchema,
+  usageReport: UsageReportSettingsSchema,
 })
