+++ src/types/settings.ts
@@ -2,11 +2,11 @@ import {
   AppSettingsSchema,
   ProxySettingsSchema,
   RecorderSettingsSchema,
-  TelemetrySettingsSchema,
+  UsageReportSettingsSchema,
 } from '@/schemas/appSettings'
 import { z } from 'zod'
 
 export type AppSettings = z.infer<typeof AppSettingsSchema>
 export type ProxySettings = z.infer<typeof ProxySettingsSchema>
 export type RecorderSettings = z.infer<typeof RecorderSettingsSchema>
-export type TelemetrySettings = z.infer<typeof TelemetrySettingsSchema>
+export type UsageReportSettings = z.infer<typeof UsageReportSettingsSchema>
